requirements.txt
pymongo==3.10.1
python-telegram-bot==12.6.1
boto==2.49.0